\name{AJL-package}
\alias{AJL-package}
\alias{AJL}
\docType{package}
\title{
\packageTitle{AJL}
}
\description{
\packageDescription{AJL}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{AJL}
\packageIndices{AJL}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{AJL}

Maintainer: \packageMaintainer{AJL}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
