% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateRelease.R
\name{simulateRelease}
\alias{simulateRelease}
\title{Simulate}
\usage{
simulateRelease(tau = 1, release.rate, n, num.obs.nonwhite, a, b, mu, sigma,
  k = 0.32, theta = 77, haphazard = FALSE, plot = TRUE)
}

